+++ scripts/check/utils/typescript.ts
